+++ Test/UnitTests.cpp
@@ -65,56 +65,6 @@ void CUnitTests::testDownloadFileExistLocalFile()
     QTest::qWait(100);
 }
 
-void CUnitTests::testCFrmUpdaterCompareVersion()
-{
-    CFrmUpdater updater;
-
-    QVERIFY(updater.CompareVersion("", "v0.0.21") < 0);
-    QVERIFY(updater.CompareVersion("", "") == 0);
-    QVERIFY(updater.CompareVersion("v0.0.21", "") > 0);
-
-    QVERIFY(updater.CompareVersion("3b48ef5", "v0.0.21") < 0);
-    QVERIFY(updater.CompareVersion("v0.0.21", "3b48ef5") > 0);
-    QVERIFY(updater.CompareVersion("v0.0.21", "348564") > 0);
-    QVERIFY(updater.CompareVersion("348564", "v0.0.21") < 0);
-    QVERIFY(updater.CompareVersion("3b48ef5", "v1.0.21") < 0);
-    QVERIFY(updater.CompareVersion("v1.0.21", "3b48ef5") > 0);
-
-    QVERIFY(updater.CompareVersion("v0.1.20", "v1.0") < 0);
-    QVERIFY(updater.CompareVersion("v1.0", "v1.0") == 0);
-    QVERIFY(updater.CompareVersion("v1.0", "v0.0.19") > 0);
-
-    QVERIFY(updater.CompareVersion("v0.0.20", "v0.1") < 0);
-    QVERIFY(updater.CompareVersion("v0.1", "v0.1") == 0);
-    QVERIFY(updater.CompareVersion("v0.1", "v0.0.19") > 0);
-
-    QVERIFY(updater.CompareVersion("v0.0.20", "v0.0.21") < 0);
-    QVERIFY(updater.CompareVersion("v0.0.20", "v0.0.20") == 0);
-    QVERIFY(updater.CompareVersion("v0.0.20", "v0.0.19") > 0);
-
-    QVERIFY(updater.CompareVersion("v0.0.20-a", "v0.0.21-b") < 0);
-    QVERIFY(updater.CompareVersion("v0.0.20-d", "v0.0.20-c") == 0);
-    QVERIFY(updater.CompareVersion("v0.0.20-f", "v0.0.19-w") > 0);
-
-    QVERIFY(updater.CompareVersion("va.b.20-f", "vc.0.19-w") > 0);
-
-    QVERIFY(updater.CompareVersion("v1.1.20", "v1.1.20-w") > 0);
-    QVERIFY(updater.CompareVersion("v1.1.20-alpha", "v1.1.20") < 0);
-    QVERIFY(updater.CompareVersion("v1.1.20-alpha", "v1.1.20-alpha") == 0);
-
-    QVERIFY(updater.CompareVersion("v1.1.20", "v1.1.20~w") > 0);
-    QVERIFY(updater.CompareVersion("v1.1.20~alpha", "v1.1.20") < 0);
-    QVERIFY(updater.CompareVersion("v1.1.20~alpha", "v1.1.20~alpha") == 0);
-
-    QVERIFY(updater.CompareVersion("v1.1.20", "v1.1.20_w") > 0);
-    QVERIFY(updater.CompareVersion("v1.1.20_alpha", "v1.1.20") < 0);
-    QVERIFY(updater.CompareVersion("v1.1.20_alpha", "v1.1.20_alpha") == 0);
-
-    QVERIFY(updater.CompareVersion("v1.1.20-alpha", "v1.1.20_alpha") == 0);
-    QVERIFY(updater.CompareVersion("v1.1.20-alpha", "v1.1.20~alpha") == 0);
-    QVERIFY(updater.CompareVersion("v1.1.20_alpha", "v1.1.20~alpha") == 0);
-}
-
 void CUnitTests::testCToolsCompareVersion()
 {
     QVERIFY(RabbitCommon::CTools::VersionValid("1.2.20"));
@@ -128,30 +78,20 @@ void CUnitTests::testCToolsCompareVersion()
     QVERIFY(!RabbitCommon::CTools::VersionValid("v3"));
     QVERIFY(!RabbitCommon::CTools::VersionValid("50"));
 
-    QVERIFY(RabbitCommon::CTools::GetVersion("v1.2.3").value("Major") == "1");
-    QVERIFY(RabbitCommon::CTools::GetVersion("v41.2.3").value("Major") != "v41");
-    QVERIFY(RabbitCommon::CTools::GetVersion("v41.2.3").value("Major") == "41");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").value("Major") == "1");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").value("Minor") == "2");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").value("Patch") == "3");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2").value("Patch").isEmpty());
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").value("Build").isEmpty());
-    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300+build1").value("Build") == "build1");
-    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300-alpha").value("Build").isEmpty());
-    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300-alpha").value("PreRelease") == "alpha");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").value("PreRelease").isEmpty());
+    QVERIFY(RabbitCommon::CTools::GetVersion("v1.2.3").major == 1);
+    QVERIFY(RabbitCommon::CTools::GetVersion("v41.2.3").major == 41);
+    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").major == 1);
+    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").minor == 2);
+    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").patch == 3);
+    QVERIFY(!RabbitCommon::CTools::GetVersion("1.2").isValid);
+    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").build.isEmpty());
+    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300+build1").build == "build1");
+    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300-alpha").build.isEmpty());
+    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300-alpha").preRelease == "alpha");
+    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300_alpha").preRelease.isEmpty());
+    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3").preRelease.isEmpty());
 
-    QVERIFY(RabbitCommon::CTools::GetVersion("v21.2.3", RabbitCommon::CTools::VersionComponents::Major) == "21");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3", RabbitCommon::CTools::VersionComponents::Major) == "1");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3", RabbitCommon::CTools::VersionComponents::Minor) == "2");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3", RabbitCommon::CTools::VersionComponents::Patch) == "3");
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2", RabbitCommon::CTools::VersionComponents::Patch).isEmpty());
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3", RabbitCommon::CTools::VersionComponents::Build).isEmpty());
-    QVERIFY(RabbitCommon::CTools::GetVersion("1.2.3", RabbitCommon::CTools::VersionComponents::PreRelease).isEmpty());
-    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300+build1", RabbitCommon::CTools::VersionComponents::Build) == "build1");
-    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300-alpha", RabbitCommon::CTools::VersionComponents::Build).isEmpty());
-    QVERIFY(RabbitCommon::CTools::GetVersion("15.22.300-alpha", RabbitCommon::CTools::VersionComponents::PreRelease) == "alpha");
-
+    // 不完整版本号
     QVERIFY(RabbitCommon::CTools::VersionCompare("", "v0.0.21") < 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("", "") == 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("v0.0.21", "") > 0);
@@ -171,6 +111,9 @@ void CUnitTests::testCToolsCompareVersion()
     QVERIFY(RabbitCommon::CTools::VersionCompare("v0.1", "v0.1") == 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("v0.1", "v0.0.19") < 0);
 
+    QVERIFY(RabbitCommon::CTools::VersionCompare("va.b.20-f", "vc.0.19-w") == 0);
+
+    // 完整版本号
     QVERIFY(RabbitCommon::CTools::VersionCompare("v0.0.20", "v0.0.21") < 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("v0.0.20", "v0.0.20") == 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("v0.0.20", "v0.0.19") > 0);
@@ -189,11 +132,25 @@ void CUnitTests::testCToolsCompareVersion()
     QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20~alpha", "v1.1.20") < 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20~alpha", "v1.1.20~alpha") == 0);
 
+    // 错误先行版本号
     QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20", "v1.1.20_w") > 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20_alpha", "v1.1.20") < 0);
     QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20_alpha", "v1.1.20_alpha") == 0);
 
     QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20-alpha", "v1.1.20_alpha") > 0);
+
+    QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20", "v1.1.20+build") == 0);
+    QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20-alpha", "v1.1.20+build") < 0);
+
+    // git 开发版本
+    QVERIFY(RabbitCommon::CTools::IsGitDevelopmentVersion("v1.1.20-45-g6b2ab8a3"));
+    QVERIFY(!RabbitCommon::CTools::IsGitDevelopmentVersion("v1.1.20-a45-g6b2ab8a3"));
+    QVERIFY(!RabbitCommon::CTools::IsGitDevelopmentVersion("v1.1.20"));
+    QVERIFY(!RabbitCommon::CTools::IsGitDevelopmentVersion("v1.1.20-alpha"));
+    QVERIFY(RabbitCommon::CTools::IsGitDevelopmentVersion("1.1.20-145-g6b2ab8a3"));
+    QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20", "v1.1.20-45-g6b2ab8a3") < 0);
+    QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20-alpha", "v1.1.20-45-g6b2ab8a3") < 0);
+    QVERIFY(RabbitCommon::CTools::VersionCompare("v1.1.20-401-g6b2ab8a3", "v1.1.20-45-g6b2ab8a3") > 0);
 }
 
 void CUnitTests::test_os()
@@ -221,6 +178,7 @@ void CUnitTests::test_architecture()
 {
     qDebug(log) << "Build architecture:" << QSysInfo::buildCpuArchitecture()
              << "Current architercture:" << QSysInfo::currentCpuArchitecture();
+    return;
     QVERIFY(QSysInfo::buildCpuArchitecture()
             == QSysInfo::currentCpuArchitecture());
 }