+++ ChangeLog.md
@@ -2,8 +2,7 @@
 - Add threa pool
 - Add file transfer
 - CTools:
-  - Add static QString GetVersion(const QString &szVersion, VersionComponents component);
-  - Add static QMap<QString, QString> GetVersion(const QString& szVersion);
+  - Add static VersionInfo GetVersion(const QString& szVersion);
   - Add static bool VersionValid(const QString& szVersion);
   - Add static int VersionCompare(const QString &ver1, const QString &ver2);
 - Add set applicaiont font